requirements.txt
streamlit
gspread
google-auth
pandas
datetime
matplotlib
plotly
time